# docs/openapi.yml
openapi: 3.0.3
info:
  title: Final Project API
  version: "1.0.0"
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        "201":
          description: Created
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: OK
  /auth/logout:
    post:
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
  /me:
    get:
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
  /courses:
    get:
      summary: List courses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
    post:
      summary: Create course
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /courses/{id}:
    get:
      summary: Get course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Update course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    delete:
      summary: Delete course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /courses/{id}/modules:
    get:
      summary: List modules by course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      summary: Create module for course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
  /courses/{id}/modules/{moduleId}:
    get:
      summary: Get module
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Update module
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    delete:
      summary: Delete module
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /modules/{id}/contents:
    get:
      summary: List contents by module
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      summary: Create content for module
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
  /modules/{id}/contents/{contentId}:
    get:
      summary: Get content
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: contentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Update content
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: contentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    delete:
      summary: Delete content
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: contentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /exams:
    get:
      summary: List exams
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
    post:
      summary: Create exam
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /exams/{id}:
    get:
      summary: Get exam
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Update exam
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    delete:
      summary: Delete exam
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /question-bank:
    get:
      summary: List question banks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
    post:
      summary: Create question bank
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /question-bank/{id}:
    get:
      summary: Get question bank
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Update question bank
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    delete:
      summary: Delete question bank
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /questions:
    get:
      summary: List questions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
    post:
      summary: Create question
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /questions/{id}:
    get:
      summary: Get question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Update question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    delete:
      summary: Delete question
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /exams/{id}/start:
    post:
      summary: Start exam attempt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
  /attempts/{id}/answer:
    post:
      summary: Submit an answer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
  /attempts/{id}/submit:
    post:
      summary: Submit attempt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /attempts/{id}/report:
    get:
      summary: Attempt report
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK